#12 move unit tests to own folder
